+++ src/utils/buttons.js
@@ -1,4 +1,38 @@
-export const buttonsCreator = (handleOperation, handlePress) => {
+export const buttonsCreator = (userText, setUserText, setCalcText) => {
+  const handlePress = number => {
+    if (number === '=') {
+      return calculation();
+    }
+    setUserText(userText + number);
+  };
+
+  const calculation = () => {
+    // eslint-disable-next-line no-eval
+    setCalcText(eval(userText));
+  };
+
+  const handleOperation = op => {
+    if (op === 'Clear') {
+      setUserText('');
+      setCalcText(0);
+      return;
+    }
+
+    if (op === 'Del') {
+      setUserText(userText.toString().substring(0, userText.length - 1));
+      return;
+    }
+
+    if (
+      ['Del', 'Clear', '+', '-', '*', '/'].includes(
+        userText.toString().split('').pop(),
+      )
+    ) {
+      return;
+    }
+    setUserText(userText + op);
+  };
+
   const buttons = [
     {
       type: 'operations',
